era5: add zfill option to zero fill rasters
When exact_extract processes data, it treats nodata values specially.
If nodata values are present in aggregation, it returns NaN. rasterio
also behaves similarly. This adds a boolean zfill option to
as_rasterio() that when set, returns a filled version of the underlying
numpy.ma.MaskedArray, with the fill set to 0 and nodata option deleted
from the raster profile.

For the specific use case of ERA5 population weighted aggregation,
setting zfill=True makes sense as we would expect unmapped areas to not
have any population.