Put in more checks for EPSG:4326
Currently geoglue is optimised for, but not exclusively for working with
grid data in lonlat (EPSG:4326) format. We should not inadvertently
assume EPSG:4326 without checking. This commit fixes this issue by
raising ValueError in appropriate places and prompting the user to
reproject before using functions such as MemoryRaster.griddes or
resample().