feat(zonal_stats): add area weighted sum
Handle area weighted sum special case. Performing weighted_sum in
exact_extract with an area coverage weight uses the formula
sum(x . c . w) where x is the raster, c is the coverage weight
(in our case, spherical area approximation in km^2) and w is the weight,
here population. This leads to very large values, especially if we use
area in m^2 and is also the wrong unit. This commit adds a special
'area_weighted_sum' operation to divide this by sum(c) which is the
total area of the polygon.